build(docker): remove USER directive for 65534 (nobody)
Binding to port 67, a privileged port, requires us to be root. There is
no way to specify which port to bind to in the CoreDHCP command line
(e.g. via a flag) and it must be done in the config. Thus, it is likely
safe to assume the default port 67 will be used in CoreDHCP
configurations.